Fixes after int-audit (#154)
* updated import ordering, interfaces first

* renamed UnacceptableTrancheIndex to UnacceptableTranche

* removed redundant trancheCount variable

* renamed tranche data to bond tranches

* renamed getTrancheIndex to indexOf

* updated indexOf return type to uint8

* renamed time to maturity to seconds to maturity

* removed unused duration method

* using uint8 for tranche iterator

* using muldiv for collaterlization math, renamed variable trancheValue to trancheAmt

* removed redundant variable and skipping fee computation is there is no fee

* initialzing allowDebase in the strategy contract initializer

* marked  external

* perp natspec comment update

* updated min lib fn usage style in perp

* calling the erc20 initializer

* updated sol version in vault

* removed invalid config check

* marked pause and unpause external

* updated muldiv parameter ordering

* renamed totalAssets to totalValue

* injection minimized implementation for

* consistant parameter ordering for  check

* renamed vaultTokenIdx to vaultTrancheIdx

* minor optimization, updating variable onlywhen iterator is updated

* gas optimization, syncing earned and underlying balances at the end of loop

* cleaned up sync asset methods, removed additional call and return value, replaced with inline replication

* marked interface earnedCount method as view

* floded redeem methods into recover, and checking for balance before executing redeem

* vault max approval on tranche and rollover

* rolled back index update change, as it results in access out of bonds

* test expecting uint8 after return type update

* ran linter

* comment update

* updated spec

+++ spot-contracts/tasks/deploy/perp.ts
@@ -111,7 +111,7 @@ task("deploy:DiscountStrategy:setDiscount")
     const discountStrategy = await hre.ethers.getContractAt("TrancheClassDiscountStrategy", discountStrategyAddress);
     const abiCoder = new hre.ethers.utils.AbiCoder();
     const hash = hre.ethers.utils.keccak256(
-      abiCoder.encode(["address", "uint256[]", "uint256"], [collateralTokenAddress, trancheRatios, trancheIndex]),
+      abiCoder.encode(["address", "uint256[]", "uint8"], [collateralTokenAddress, trancheRatios, trancheIndex]),
     );
     const tx = await discountStrategy.updateDefinedDiscount(
       hash,
